LayoutTests/fast/images/img-dimensions-styled.html

<!DOCTYPE html>
<title>HTMLImageElement.width/height returns styled dimensions.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
    img {
        width: 200px;
        height: 200px;
    }
</style>
<img src="resources/test-load.jpg" width="100" height="100">
<script>
test(function() {
    var image = document.querySelector("img");
    assert_equals(image.width, 200);
    assert_equals(image.height, 200);
    var rect = image.getBoundingClientRect();
    assert_equals(rect.width, 200);
    assert_equals(rect.height, 200);
});
</script>
